feat: add installation script
The script is pretty comprehensive and should cover most shells. Poetry
and python are not really a great fit for making a CLI tool. This is the
best I could come up with (a custom shell script writer that runs squire
for you). Might be some problems with some shells and still requires one
to have bash installed